+++ NQueen.Kernel/Utilities/ProgressSettings.cs
@@ -4,7 +4,7 @@ public static class ProgressSettings
 {
     public const double StartProgressValue = 0;
 
-    public static int SolutionCountPerUpdate(byte boardSize) =>
+    public static int SolutionCountPerUpdate(int boardSize) =>
         boardSize <= BoardSettings.SmallBoardSizeForUniqueSolutions
         ? 5
         : boardSize <= BoardSettings.MediumBoardSizeForUniqueSolutions
